[AVL]: 将平衡二叉树内的结点索引类型改为int64, 避免id的冲突概率

+++ src/incl/avl_tree.h
@@ -57,7 +57,7 @@ typedef int (*avl_cmp_cb_t)(const void *key, const void *data);
 /* 节点结构 */
 typedef struct _node_t
 {
-    unsigned int idx;           /* 索引值: 该值有key_cb生成 可能不唯一 */
+    int64_t idx;                /* 索引值: 该值有key_cb生成 可能不唯一 */
 
     int bf;                     /* 平衡因子 */
